requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
ultralytics==8.0.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1